room 수정 삭제 등록 완료 이미지부분 추가해야함

+++ travelerPro/src/main/java/com/room/RoomDTO.java
@@ -78,7 +78,7 @@ public class RoomDTO {
 	public long getCompanyNum() {
 		return companyNum;
 	}
-	public void setCompanyNum(int companyNum) {
+	public void setCompanyNum(long companyNum) {
 		this.companyNum = companyNum;
 	}
 	public int getHeadCount() {
